Housekeeping - resolve my rubocop issues
Some issues I'm choosing not to resolve in order to keep within the convention of a single method, written as provided, in which to solve the problem.

This means they'll often be too long and complex, and unfortunately come with some dubious naming.

The rest though, this commit cleans up so my solutions stay on the rubyv straight and narrow!